+++ util/src/platform/win/backtrace_win.c
@@ -2,7 +2,7 @@
 #include <stdio.h>
 
 char**
-get_backtrace(intptr_t* size )
+get_backtrace(uint32_t* size )
 {
     /* TODO */
     fprintf(stderr, "backtraces not implemented on windows\n");
